+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Remove the explicit dependency on the RUM extension (it's now implicit on the .so file). Flip to documentdb_extended_rum for PG18+ *[Feature]*
 * Use the appropriate GUC for the user_crud_commands.sql *[Bugfix]* (#319)
 * Provide Rust dev environment in devcontainer *[Feature]*
+* Add extension that adds a gateway host that's run as a postgres background worker *[Feature]*
 
 ### documentdb v0.107-0 (Unreleased) ###
 * Support sort by _id against the _id index using the enableIndexOrderbyPushdown flag *[Feature]*.